# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvoiceFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-profile:
    get:
      operationId: getBusinessProfile
      tags: [business-profiles]
      summary: Get business profile
      responses:
        "200":
          description: Business profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBusinessProfile
      tags: [business-profiles]
      summary: Create business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessProfileBody"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
    patch:
      operationId: updateBusinessProfile
      tags: [business-profiles]
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Customers
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Products
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        shopName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        accountHolder:
          type: ["string", "null"]
        invoicePrefix:
          type: string
        invoiceStartNumber:
          type: integer
        defaultTaxPercent:
          type: number
        defaultNotes:
          type: ["string", "null"]
        defaultPaymentTerms:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shopName
        - ownerName
        - email
        - phone
        - invoicePrefix
        - invoiceStartNumber
        - defaultTaxPercent
        - createdAt
        - updatedAt
    CreateBusinessProfileBody:
      type: object
      properties:
        shopName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        gstin:
          type: string
        website:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        ifscCode:
          type: string
        accountHolder:
          type: string
        invoicePrefix:
          type: string
        defaultTaxPercent:
          type: number
        defaultNotes:
          type: string
        defaultPaymentTerms:
          type: string
      required:
        - shopName
        - ownerName
        - email
        - phone
    UpdateBusinessProfileBody:
      type: object
      properties:
        shopName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        gstin:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        ifscCode:
          type: string
        accountHolder:
          type: string
        invoicePrefix:
          type: string
        defaultTaxPercent:
          type: number
        defaultNotes:
          type: string
        defaultPaymentTerms:
          type: string
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - createdAt
    CustomerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        totalInvoices:
          type: integer
        totalBilled:
          type: number
        totalPaid:
          type: number
        outstanding:
          type: number
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceListItem"
      required:
        - id
        - name
        - phone
        - createdAt
        - totalInvoices
        - totalBilled
        - totalPaid
        - outstanding
        - invoices
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        gstin:
          type: string
      required:
        - name
        - phone
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        gstin:
          type: string
...
    InvoiceListItem:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        total:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - invoiceDate
        - status
        - total
        - createdAt
...